+++ src/encoded/static/components/top_hits/constants.js
@@ -1,12 +1,3 @@
-import {
-    Query as TopHitsQuery,
-    CollectionsQuery,
-} from './query';
-import Results, {
-    CollectionResults
-} from './results';
-
-
 export const TOP_HITS_URL = '/top-hits-raw/';
 
 export const TOP_HITS_PARAMS = (
@@ -112,11 +103,6 @@ export const COLLECTIONS = [
         '@type': ['DataCollection'],
     }
 
-]
-
-export const COLLECTIONS_KEY = 'DataCollection'
-
-export const queries = [
-    ['dataCollections', CollectionsQuery, CollectionResults, 'Data collections'],
-    ['topHits', TopHitsQuery, Results, 'Top results by type'],
 ];
+
+export const COLLECTIONS_KEY = 'DataCollection';
