+++ src/encoded/static/components/top_hits/multi/constants.js
@@ -1,0 +1,12 @@
+import {
+    Query as TopHitsQuery,
+    CollectionsQuery,
+} from '../query';
+import TopHitsResults from '../results';
+import CollectionsResults from './results';
+
+
+export const queries = [
+    ['dataCollections', CollectionsQuery, CollectionsResults, 'Data collections'],
+    ['topHits', TopHitsQuery, TopHitsResults, 'Top results by type'],
+];
